app/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-exporter-google-cloud==0.39b0
opentelemetry-exporter-cloud-monitoring==0.39b0